Feature/scratch card game

+++ src/main/kotlin/casino/core/payment/TestPaymentService.kt
@@ -3,7 +3,7 @@ package main.kotlin.casino.core.payment
 import main.kotlin.casino.core.domain.Player
 
 class TestPaymentService: IPaymentService {
-    override fun addFunds(player: Player, amount: Double) {
+    override fun addFunds(player: Player, amount: Float) {
         TODO("Not yet implemented")
     }
 }
